Allow custom headers to be used
Expose a headers hash when making API calls. This allows custom headers to be sent
to the origin server. All provided headers are sent to the server.

If the Accept header is not supplied then application/json is used as a default.

eg

Model.get('path', headers: { cache_control: 'no-cache'})

Will make the HTTP request with Cache-Control: no-cache header