Make API user agent explicit
We want to know how our current API's are used.
Part of this is knowing (tracking) which API endpoints are used by which apps/services - and more importantly which current API endpoints are just not being used!
To do this we can send a custom User-Agent along with each request that identifies the originator of the request.
The intention is to send this custom User-Agent to honeycomb.