Deprecate `:dig` and `:[]` first.
We recently removed these methods to prevent inner-object access of non existent attributes via [] or dig (#110).
However, the change was disruptive for everyones workflow and requires a lot of focus to migrate the api_clients, tests and services.

So I'm releasing a new _minor_ version to bring back this methods but deprecating the methods.
Then, when we want to move forward with a "stricter" version, we can configure the `disallowed_behavior` in Rails first to be ready for the migration.

This PR won't break any behaviour, in fact, it will unblock some dependabot PRs.