Use custom ActiveSupport::Deprecator
Creates a custom ActiveSupport::Deprecator to deprecate methods on
`StrictOpenStruct`.

This will ensure that deprecation notices are attributed to
`restful_resource` and not the default of Rails@6.2

This means that projects with custom `ActiveSupport::Notification`
subscribers (such as all internal carwow projects, via our honeycomb
subscribers) for deprecations will be able to differentiate them from
normal Rails 6.2 deprecation notices.